EXPERIMENTAL: kthread: preempt wait_for_completion
This commit attempts to use an interruptible wait_for_completion method
called wait_for_completion_interruptible. My hopes for this commit is to
reduce the blocking nature of these waits when they are parked or
unparked.

This commit is hardly tested; all I know is that the device boots
without issue with this commit applied. Further research and testing
will be done.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>